Fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38167 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/snapshot.c
@@ -585,7 +585,7 @@ snapshot_module_t *snapshot_module_create(snapshot_t *s, const char *name, uint8
         return NULL;
     }
 
-    m->size = ftell(s->file) - m->offset;
+    m->size = (uint32_t)(ftell(s->file) - m->offset);
     m->size_offset = ftell(s->file) - sizeof(uint32_t);
 
     return m;
